Migrated DoublesBattle message filter.
